add pagination, categorization, cart

+++ apps/products/models.py
@@ -7,6 +7,6 @@ class Product(models.Model):
   desc = models.TextField()
   category = models.CharField(max_length=50)
   image = models.CharField(max_length=100)
-  price = models.DecimalField(max_digits=10, decimal_places=2)
+  price = models.IntegerField()
   created_at = models.DateTimeField(auto_now_add = True)
   updated_at = models.DateTimeField(auto_now_add = True)
